feat(assoc): Throw if already initialized
